mkt-mgs-handler/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-firestore
google-api-python-client
